requirements.txt
asgiref==3.11.1
dj-rest-auth==7.2.0
Django==5.2.13
django-allauth==65.16.1
django-ckeditor==6.7.3
django-extensions==4.1
django-js-asset==3.1.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
gunicorn==25.3.0
inflection==0.5.1
packaging==26.2
pillow==12.2.0
psycopg2-binary==2.9.12
PyJWT==2.12.1
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
whitenoise==6.12.0
